Change second x to a y
Covariance is between two random variables but the formula taken from Wikipedia provides a definition for variance of a single random variable. I've changed the second x in the formulae to a y to make it clear that we are considering how two random variables covary.
I suppose Wikipedia will also need changing, but I've never done that before.